@@ -4,13 +4,13 @@ version: 0.1.0+1
 publish_to: none
 
 environment:
-  sdk: ">=2.19.0 <3.0.0"
+  sdk: ">=3.0.0-134.0.dev <4.0.0"
 
 dev_dependencies:
-  build_runner: ^2.3.3
+  build_runner: ^2.4.1
   json_serializable: ^6.6.1
   mocktail: ^0.3.0
-  test: ^1.19.2
+  test: ^1.24.1
   very_good_analysis: ^4.0.0
 
 dependencies:
@@ -18,3 +18,11 @@ dependencies:
   game_script_machine:
     path: ../game_script_machine
   json_annotation: ^4.8.0
+
+# There seems to be a breaking change in the current beta
+# that makes the resolved verion of pub_semver that is
+# a transitive dependency from build_runner to break.
+#
+# Overriding to the latest one in the mean time.
+dependency_overrides:
+  pub_semver: ^2.1.3
