@@ -13,7 +13,6 @@ class LeaderboardRepository {
         _blacklistDocumentId = blacklistDocumentId;
 
   final DbClient _dbClient;
-
   final String _blacklistDocumentId;
 
   /// Retrieves the blacklist for player initials.
@@ -47,5 +46,49 @@ class LeaderboardRepository {
       'id': results.first.id,
       ...results.first.data,
     });
+  }
+
+  /// Retrieves the top score cards with the highest total wins.
+  Future<List<ScoreCard>> getScoreCardsWithMostWins() async {
+    final results = await _dbClient.orderBy('score_cards', 'wins');
+
+    return results
+        .map(
+          (e) => ScoreCard.fromJson({
+            'id': e.id,
+            ...e.data,
+          }),
+        )
+        .toList();
+  }
+
+  /// Retrieves the top score cards with the longest streak.
+  Future<List<ScoreCard>> getScoreCardsWithLongestStreak() async {
+    final results = await _dbClient.orderBy('score_cards', 'longestStreak');
+
+    return results
+        .map(
+          (e) => ScoreCard.fromJson({
+            'id': e.id,
+            ...e.data,
+          }),
+        )
+        .toList();
+  }
+
+  /// Adds the initials to the score card with the given [scoreCardId].
+  Future<void> addInitialsToScoreCard({
+    required String scoreCardId,
+    required String initials,
+  }) async {
+    await _dbClient.update(
+      'score_cards',
+      DbEntityRecord(
+        id: scoreCardId,
+        data: {
+          'initials': initials,
+        },
+      ),
+    );
   }
 }
