@@ -1,5 +1,6 @@
 import 'package:api_client/api_client.dart';
 import 'package:authentication_repository/authentication_repository.dart';
+import 'package:connection_repository/connection_repository.dart';
 import 'package:equatable/equatable.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
@@ -7,13 +8,11 @@ import 'package:game_domain/game_domain.dart';
 import 'package:go_router/go_router.dart';
 import 'package:match_maker_repository/match_maker_repository.dart';
 import 'package:top_dash/game/game.dart';
-import 'package:web_socket_client/web_socket_client.dart';
 
 class GamePage extends StatelessWidget {
   const GamePage({
     required this.matchId,
     required this.isHost,
-    required this.matchConnection,
     super.key,
   });
 
@@ -24,13 +23,11 @@ class GamePage extends StatelessWidget {
       key: const Key('game'),
       matchId: data?.matchId ?? '',
       isHost: data?.isHost ?? false,
-      matchConnection: data?.matchConnection,
     );
   }
 
   final String matchId;
   final bool isHost;
-  final WebSocket? matchConnection;
 
   @override
   Widget build(BuildContext context) {
@@ -38,15 +35,16 @@ class GamePage extends StatelessWidget {
       create: (context) {
         final gameResource = context.read<GameResource>();
         final matchMakerRepository = context.read<MatchMakerRepository>();
+        final connectionRepository = context.read<ConnectionRepository>();
         final matchSolver = context.read<MatchSolver>();
         final user = context.read<User>();
         return GameBloc(
           gameResource: gameResource,
           matchMakerRepository: matchMakerRepository,
+          connectionRepository: connectionRepository,
           matchSolver: matchSolver,
           user: user,
           isHost: isHost,
-          matchConnection: matchConnection,
         )..add(MatchRequested(matchId));
       },
       child: const GameView(),
@@ -58,13 +56,11 @@ class GamePageData extends Equatable {
   const GamePageData({
     required this.isHost,
     required this.matchId,
-    required this.matchConnection,
   });
 
   final bool isHost;
   final String? matchId;
-  final WebSocket? matchConnection;
 
   @override
-  List<Object?> get props => [isHost, matchId, matchConnection];
+  List<Object?> get props => [isHost, matchId];
 }
