@@ -13,10 +13,10 @@ class GameSummaryView extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return Scaffold(
+    return const Scaffold(
       backgroundColor: TopDashColors.seedWhite,
       body: Stack(
-        children: const [
+        children: [
           Align(
             child: _MatchSummaryScreenView(key: Key('match_summary_view')),
           ),
@@ -51,9 +51,9 @@ class PortraitSummaryView extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return Column(
+    return const Column(
       mainAxisAlignment: MainAxisAlignment.center,
-      children: const [
+      children: [
         _ResultView(key: Key('game_summary_result_view')),
         SizedBox(height: TopDashSpacing.xxlg),
         _CardsView(),
@@ -67,9 +67,9 @@ class LandscapeSummaryView extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return Row(
+    return const Row(
       mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-      children: const [
+      children: [
         _ResultView(key: Key('game_summary_result_view')),
         _CardsView(),
       ],
@@ -189,6 +189,9 @@ class GameSummaryFooter extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     final l10n = context.l10n;
+    final bloc = context.read<GameBloc>();
+    final state = bloc.state as MatchLoadedState;
+    final playerScoreCard = state.playerScoreCard;
 
     return ColoredBox(
       color: TopDashColors.seedWhite,
@@ -208,7 +211,12 @@ class GameSummaryFooter extends StatelessWidget {
               onPressed: () => QuitGameDialog.show(
                 context,
                 onConfirm: () => _routerNeglectCall(context, () {
-                  GoRouter.of(context).go('/');
+                  if (playerScoreCard.initials != null) {
+                    GoRouter.of(context).go('/');
+                  } else {
+                    GoRouter.of(context).pop();
+                    bloc.add(const LeaderboardEntryRequested());
+                  }
                 }),
                 onCancel: GoRouter.of(context).pop,
               ),
