@@ -2,27 +2,18 @@ import 'package:flutter/material.dart';
 import 'package:top_dash_ui/top_dash_ui.dart';
 
 class LeaderboardPlayers extends StatelessWidget {
-  const LeaderboardPlayers({super.key});
+  const LeaderboardPlayers({
+    required this.players,
+    super.key,
+  });
 
-  // TODO(willhlas): replace with real data once available.
-  static const _playerList = [
-    _Player(index: 0, initials: 'AAA', wins: 220),
-    _Player(index: 1, initials: 'BBB', wins: 218),
-    _Player(index: 2, initials: 'CCC', wins: 172),
-    _Player(index: 3, initials: 'DDD', wins: 154),
-    _Player(index: 4, initials: 'EEE', wins: 152),
-    _Player(index: 5, initials: 'FFF', wins: 144),
-    _Player(index: 6, initials: 'GGG', wins: 127),
-    _Player(index: 7, initials: 'HHH', wins: 99),
-    _Player(index: 8, initials: 'III', wins: 97),
-    _Player(index: 9, initials: 'JJJ', wins: 81),
-  ];
+  final List<LeaderboardPlayer> players;
 
   @override
   Widget build(BuildContext context) {
     return Column(
       children: [
-        for (final player in _playerList) ...[
+        for (final player in players) ...[
           player,
           const SizedBox(height: TopDashSpacing.md),
         ],
@@ -31,16 +22,17 @@ class LeaderboardPlayers extends StatelessWidget {
   }
 }
 
-class _Player extends StatelessWidget {
-  const _Player({
+class LeaderboardPlayer extends StatelessWidget {
+  const LeaderboardPlayer({
     required this.index,
     required this.initials,
-    required this.wins,
+    required this.value,
+    super.key,
   });
 
-  final String initials;
-  final int wins;
   final int index;
+  final String initials;
+  final int value;
 
   @override
   Widget build(BuildContext context) {
@@ -78,7 +70,7 @@ class _Player extends StatelessWidget {
           ],
         ),
         Text(
-          wins.toString(),
+          value.toString(),
           style: TopDashTextStyles.buttonLG,
         ),
       ],
