@@ -1,11 +1,12 @@
 import 'dart:async';
 
 import 'package:api_client/api_client.dart';
+import 'package:connection_repository/connection_repository.dart';
 import 'package:equatable/equatable.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
+import 'package:game_domain/game_domain.dart';
 import 'package:match_maker_repository/match_maker_repository.dart';
-import 'package:web_socket_client/web_socket_client.dart';
 
 part 'match_making_event.dart';
 part 'match_making_state.dart';
@@ -13,11 +14,13 @@ part 'match_making_state.dart';
 class MatchMakingBloc extends Bloc<MatchMakingEvent, MatchMakingState> {
   MatchMakingBloc({
     required MatchMakerRepository matchMakerRepository,
+    required ConnectionRepository connectionRepository,
     required GameResource gameResource,
     required this.cardIds,
     this.hostWaitTime = defaultHostWaitTime,
     this.guestWaitingTimeout = 30,
   })  : _matchMakerRepository = matchMakerRepository,
+        _connectionRepository = connectionRepository,
         _gameResource = gameResource,
         super(const MatchMakingState.initial()) {
     on<MatchRequested>(_onMatchRequested);
@@ -28,11 +31,13 @@ class MatchMakingBloc extends Bloc<MatchMakingEvent, MatchMakingState> {
   @visibleForTesting
   MatchMakingBloc.test({
     required MatchMakerRepository matchMakerRepository,
+    required ConnectionRepository connectionRepository,
     required GameResource gameResource,
     required this.cardIds,
     this.hostWaitTime = defaultHostWaitTime,
     this.guestWaitingTimeout = 1,
   })  : _matchMakerRepository = matchMakerRepository,
+        _connectionRepository = connectionRepository,
         _gameResource = gameResource,
         super(const MatchMakingState.initial()) {
     on<MatchRequested>(_onMatchRequested);
@@ -44,10 +49,25 @@ class MatchMakingBloc extends Bloc<MatchMakingEvent, MatchMakingState> {
   final GameResource _gameResource;
   final List<String> cardIds;
   final int guestWaitingTimeout;
-  WebSocket? _matchConnection;
+  final ConnectionRepository _connectionRepository;
 
   static const defaultHostWaitTime = Duration(seconds: 4);
   final Duration hostWaitTime;
+
+  Future<void> _connectToMatch({
+    required String matchId,
+    required bool isHost,
+  }) async {
+    _connectionRepository.send(
+      WebSocketMessage.matchJoined(
+        matchId: matchId,
+        isHost: isHost,
+      ),
+    );
+    await _connectionRepository.messages.firstWhere(
+      (message) => message.messageType == MessageType.matchJoined,
+    );
+  }
 
   Future<void> _onMatchRequested(
     MatchRequested event,
@@ -57,25 +77,21 @@ class MatchMakingBloc extends Bloc<MatchMakingEvent, MatchMakingState> {
       emit(state.copyWith(status: MatchMakingStatus.processing));
       final playerId = await _gameResource.createDeck(cardIds);
       final match = await _matchMakerRepository.findMatch(playerId);
+      final isHost = match.guest == null;
 
-      if (match.guest != null) {
-        _matchConnection = await _gameResource.connectToMatch(
-          matchId: match.id,
-          isHost: false,
-        );
+      await _connectToMatch(
+        matchId: match.id,
+        isHost: isHost,
+      );
+      if (!isHost) {
         emit(
           state.copyWith(
             match: match,
             status: MatchMakingStatus.completed,
             isHost: false,
-            matchConnection: _matchConnection,
           ),
         );
       } else {
-        _matchConnection = await _gameResource.connectToMatch(
-          matchId: match.id,
-          isHost: true,
-        );
         await _waitGuestToJoin(
           match: match,
           emit: emit,
@@ -96,7 +112,7 @@ class MatchMakingBloc extends Bloc<MatchMakingEvent, MatchMakingState> {
       final playerId = await _gameResource.createDeck(cardIds);
       final match = await _matchMakerRepository.createPrivateMatch(playerId);
 
-      _matchConnection = await _gameResource.connectToMatch(
+      await _connectToMatch(
         matchId: match.id,
         isHost: true,
       );
@@ -121,7 +137,7 @@ class MatchMakingBloc extends Bloc<MatchMakingEvent, MatchMakingState> {
         guestId: playerId,
         inviteCode: event.inviteCode,
       );
-      _matchConnection = await _gameResource.connectToMatch(
+      await _connectToMatch(
         matchId: match!.id,
         isHost: false,
       );
@@ -140,7 +156,7 @@ class MatchMakingBloc extends Bloc<MatchMakingEvent, MatchMakingState> {
   }
 
   Future<void> _waitGuestToJoin({
-    required Match match,
+    required DraftMatch match,
     required Emitter<MatchMakingState> emit,
   }) async {
     final stream = _matchMakerRepository
@@ -149,7 +165,7 @@ class MatchMakingBloc extends Bloc<MatchMakingEvent, MatchMakingState> {
 
     emit(state.copyWith(match: match));
 
-    late StreamSubscription<Match> subscription;
+    late StreamSubscription<DraftMatch> subscription;
 
     subscription = stream.listen((newMatch) {
       emit(
@@ -157,7 +173,6 @@ class MatchMakingBloc extends Bloc<MatchMakingEvent, MatchMakingState> {
           match: newMatch,
           status: MatchMakingStatus.completed,
           isHost: true,
-          matchConnection: _matchConnection,
         ),
       );
       subscription.cancel();
@@ -182,11 +197,10 @@ class MatchMakingBloc extends Bloc<MatchMakingEvent, MatchMakingState> {
               match: match.copyWithGuest(guest: 'CPU_${match.host}'),
               status: MatchMakingStatus.completed,
               isHost: true,
-              matchConnection: _matchConnection,
             ),
           );
         } catch (e) {
-          _matchConnection?.close();
+          _connectionRepository.send(const WebSocketMessage.matchLeft());
           emit(state.copyWith(status: MatchMakingStatus.timeout));
         }
       },
@@ -196,7 +210,7 @@ class MatchMakingBloc extends Bloc<MatchMakingEvent, MatchMakingState> {
   @override
   Future<void> close() {
     if (state.match?.guest == null) {
-      _matchConnection?.close();
+      _connectionRepository.send(const WebSocketMessage.matchLeft());
     }
     return super.close();
   }
