@@ -13,7 +13,6 @@ class MatchMakingState extends Equatable {
     required this.status,
     this.isHost = false,
     this.match,
-    this.matchConnection,
   });
 
   const MatchMakingState.initial()
@@ -21,22 +20,19 @@ class MatchMakingState extends Equatable {
           status: MatchMakingStatus.initial,
         );
 
-  final Match? match;
+  final DraftMatch? match;
   final MatchMakingStatus status;
   final bool isHost;
-  final WebSocket? matchConnection;
 
   MatchMakingState copyWith({
-    Match? match,
+    DraftMatch? match,
     MatchMakingStatus? status,
     bool? isHost,
-    WebSocket? matchConnection,
   }) {
     return MatchMakingState(
       match: match ?? this.match,
       status: status ?? this.status,
       isHost: isHost ?? this.isHost,
-      matchConnection: matchConnection ?? this.matchConnection,
     );
   }
 
@@ -45,6 +41,5 @@ class MatchMakingState extends Equatable {
         status,
         match,
         isHost,
-        matchConnection,
       ];
 }
