@@ -4,7 +4,6 @@ import 'dart:io';
 
 import 'package:api_client/api_client.dart';
 import 'package:game_domain/game_domain.dart';
-import 'package:web_socket_client/web_socket_client.dart';
 
 /// {@template game_resource}
 /// An api resource for interacting with the game.
@@ -13,12 +12,9 @@ class GameResource {
   /// {@macro game_resource}
   GameResource({
     required ApiClient apiClient,
-    Duration? webSocketTimeout,
-  })  : _apiClient = apiClient,
-        _webSocketTimeout = webSocketTimeout ?? const Duration(seconds: 20);
+  }) : _apiClient = apiClient;
 
   final ApiClient _apiClient;
-  final Duration _webSocketTimeout;
 
   /// Post /cards
   Future<List<Card>> generateCards(Prompt prompt) async {
@@ -184,37 +180,6 @@ class GameResource {
     } catch (e) {
       throw ApiClientError(
         'POST /matches/$matchId/decks/$deckId/cards/$cardId failed with the following message: "$e"',
-        StackTrace.current,
-      );
-    }
-  }
-
-  /// WebSocket connect to  public/matches/connect
-  Future<WebSocket> connectToMatch({
-    required String matchId,
-    required bool isHost,
-  }) async {
-    try {
-      final socket = await _apiClient.connect(
-        '/public/matches/connect',
-        queryParameters: {
-          'matchId': matchId,
-          'host': isHost.toString(),
-        },
-      );
-
-      await socket.connection.firstWhere((state) => state is Connected).timeout(
-        _webSocketTimeout,
-        onTimeout: () async {
-          socket.close();
-          throw TimeoutException('Could not connect to match');
-        },
-      );
-
-      return socket;
-    } catch (error) {
-      throw ApiClientError(
-        'websocket public/matches/connect returned with the following error: "$error"',
         StackTrace.current,
       );
     }
