@@ -3,7 +3,7 @@ import 'dart:math' as math;
 
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:flutter/foundation.dart';
-import 'package:game_domain/game_domain.dart' hide Match;
+import 'package:game_domain/game_domain.dart';
 import 'package:match_maker_repository/match_maker_repository.dart';
 import 'package:uuid/uuid.dart';
 
@@ -58,7 +58,7 @@ class MatchMakerRepository {
   late final math.Random _randomGenerator;
 
   /// Watches a match.
-  Stream<Match> watchMatch(String id) {
+  Stream<DraftMatch> watchMatch(String id) {
     return collection.doc(id).snapshots().map((snapshot) {
       final id = snapshot.id;
       final data = snapshot.data()!;
@@ -67,7 +67,7 @@ class MatchMakerRepository {
       final hostConnected = data['hostConnected'] as bool?;
       final guestConnected = data['guestConnected'] as bool?;
 
-      return Match(
+      return DraftMatch(
         id: id,
         host: host,
         guest: guest == _emptyKey || guest == _inviteKey ? null : guest,
@@ -110,7 +110,7 @@ class MatchMakerRepository {
     });
   }
 
-  Match _mapMatchQueryElement(
+  DraftMatch _mapMatchQueryElement(
     QueryDocumentSnapshot<Map<String, dynamic>> element,
   ) {
     final id = element.id;
@@ -120,7 +120,7 @@ class MatchMakerRepository {
     final guestConnected = data['guestConnected'] as bool?;
     final inviteCode = data['inviteCode'] as String?;
 
-    return Match(
+    return DraftMatch(
       id: id,
       host: host,
       hostConnected: hostConnected ?? false,
@@ -140,7 +140,7 @@ class MatchMakerRepository {
   }
 
   /// Finds a match.
-  Future<Match> findMatch(String id, {int retryNumber = 0}) async {
+  Future<DraftMatch> findMatch(String id, {int retryNumber = 0}) async {
     /// Find a match that is not full and has
     /// been updated in the last 4 seconds.
     final matchesResult = await collection
@@ -186,13 +186,13 @@ class MatchMakerRepository {
   }
 
   /// Creates a private match that can only be joined with an invitation code.
-  Future<Match> createPrivateMatch(String id) => _createMatch(
+  Future<DraftMatch> createPrivateMatch(String id) => _createMatch(
         id,
         inviteOnly: true,
       );
 
   /// Searches for and join a private match. Returns null if none is found.
-  Future<Match?> joinPrivateMatch({
+  Future<DraftMatch?> joinPrivateMatch({
     required String guestId,
     required String inviteCode,
   }) async {
@@ -219,7 +219,7 @@ class MatchMakerRepository {
     return null;
   }
 
-  Future<Match> _createMatch(String id, {bool inviteOnly = false}) async {
+  Future<DraftMatch> _createMatch(String id, {bool inviteOnly = false}) async {
     final inviteCode = inviteOnly ? _inviteCode() : null;
     final result = await collection.add({
       'host': id,
@@ -233,7 +233,7 @@ class MatchMakerRepository {
       'guestPlayedCards': const <String>[],
       'hostStartsMatch': _randomGenerator.nextBool(),
     });
-    return Match(
+    return DraftMatch(
       id: result.id,
       host: id,
       inviteCode: inviteCode,
