@@ -5,12 +5,36 @@ import 'package:bloc_test/bloc_test.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:game_domain/game_domain.dart';
 import 'package:mocktail/mocktail.dart';
+import 'package:top_dash/audio/audio_controller.dart';
 import 'package:top_dash/draft/draft.dart';
+import 'package:top_dash/gen/assets.gen.dart';
 
 class _MockGameResource extends Mock implements GameResource {}
 
+class _MockAudioController extends Mock implements AudioController {}
+
 void main() {
   group('DraftBloc', () {
+    final rareCard = Card(
+      id: '0',
+      name: '',
+      description: '',
+      image: '',
+      rarity: true,
+      power: 20,
+      suit: Suit.fire,
+    );
+
+    final commonCard = Card(
+      id: '0',
+      name: '',
+      description: '',
+      image: '',
+      rarity: false,
+      power: 20,
+      suit: Suit.fire,
+    );
+
     final cards = List.generate(
       10,
       (i) => Card(
@@ -25,23 +49,33 @@ void main() {
     );
 
     late GameResource gameResource;
+    late AudioController audioController;
 
     setUp(() {
       gameResource = _MockGameResource();
       when(() => gameResource.generateCards(Prompt()))
           .thenAnswer((_) async => cards);
+
+      audioController = _MockAudioController();
+      when(() => audioController.playSfx(any())).thenAnswer((_) {});
     });
 
     test('has the correct initial state', () {
       expect(
-        DraftBloc(gameResource: _MockGameResource()).state,
+        DraftBloc(
+          gameResource: _MockGameResource(),
+          audioController: _MockAudioController(),
+        ).state,
         equals(DraftState.initial()),
       );
     });
 
     blocTest<DraftBloc, DraftState>(
       'can request a deck',
-      build: () => DraftBloc(gameResource: gameResource),
+      build: () => DraftBloc(
+        gameResource: gameResource,
+        audioController: audioController,
+      ),
       act: (bloc) => bloc.add(DeckRequested(Prompt())),
       expect: () => [
         DraftState(
@@ -60,8 +94,52 @@ void main() {
     );
 
     blocTest<DraftBloc, DraftState>(
+      'plays the reveal sfx when the deck is loaded',
+      build: () => DraftBloc(
+        gameResource: gameResource,
+        audioController: audioController,
+      ),
+      act: (bloc) => bloc.add(DeckRequested(Prompt())),
+      verify: (_) {
+        verify(() => audioController.playSfx(Assets.sfx.reveal)).called(1);
+      },
+    );
+
+    blocTest<DraftBloc, DraftState>(
+      'plays the holo reveal sfx when the deck is loaded and the first '
+      'card is rare',
+      setUp: () {
+        final cards = List.generate(
+          10,
+          (i) => Card(
+            id: i.toString(),
+            name: '',
+            description: '',
+            image: '',
+            rarity: i == 0,
+            power: 20,
+            suit: Suit.values[i % Suit.values.length],
+          ),
+        );
+        when(() => gameResource.generateCards(Prompt()))
+            .thenAnswer((_) async => cards);
+      },
+      build: () => DraftBloc(
+        gameResource: gameResource,
+        audioController: audioController,
+      ),
+      act: (bloc) => bloc.add(DeckRequested(Prompt())),
+      verify: (_) {
+        verify(() => audioController.playSfx(Assets.sfx.holoReveal)).called(1);
+      },
+    );
+
+    blocTest<DraftBloc, DraftState>(
       'change the cards order on PreviousCard',
-      build: () => DraftBloc(gameResource: gameResource),
+      build: () => DraftBloc(
+        gameResource: gameResource,
+        audioController: audioController,
+      ),
       seed: () => DraftState(
         cards: cards,
         selectedCards: const [],
@@ -85,8 +163,31 @@ void main() {
     );
 
     blocTest<DraftBloc, DraftState>(
+      'plays the holo reveal when the previous card is rare',
+      build: () => DraftBloc(
+        gameResource: gameResource,
+        audioController: audioController,
+      ),
+      seed: () => DraftState(
+        cards: [commonCard, rareCard],
+        selectedCards: const [],
+        status: DraftStateStatus.deckLoaded,
+        firstCardOpacity: 1,
+      ),
+      act: (bloc) {
+        bloc.add(PreviousCard());
+      },
+      verify: (_) {
+        verify(() => audioController.playSfx(Assets.sfx.holoReveal)).called(1);
+      },
+    );
+
+    blocTest<DraftBloc, DraftState>(
       'change the cards order on NextCard',
-      build: () => DraftBloc(gameResource: gameResource),
+      build: () => DraftBloc(
+        gameResource: gameResource,
+        audioController: audioController,
+      ),
       seed: () => DraftState(
         cards: cards,
         selectedCards: const [],
@@ -110,8 +211,31 @@ void main() {
     );
 
     blocTest<DraftBloc, DraftState>(
+      'plays the holo reveal when the next card is rare',
+      build: () => DraftBloc(
+        gameResource: gameResource,
+        audioController: audioController,
+      ),
+      seed: () => DraftState(
+        cards: [commonCard, rareCard],
+        selectedCards: const [],
+        status: DraftStateStatus.deckLoaded,
+        firstCardOpacity: 1,
+      ),
+      act: (bloc) {
+        bloc.add(NextCard());
+      },
+      verify: (_) {
+        verify(() => audioController.playSfx(Assets.sfx.holoReveal)).called(1);
+      },
+    );
+
+    blocTest<DraftBloc, DraftState>(
       'change the cards order on CardSwiped',
-      build: () => DraftBloc(gameResource: gameResource),
+      build: () => DraftBloc(
+        gameResource: gameResource,
+        audioController: audioController,
+      ),
       seed: () => DraftState(
         cards: cards,
         selectedCards: const [],
@@ -133,10 +257,32 @@ void main() {
         ),
       ],
     );
+    blocTest<DraftBloc, DraftState>(
+      'plays the holo reveal after a swipe and when the next card is rare',
+      build: () => DraftBloc(
+        gameResource: gameResource,
+        audioController: audioController,
+      ),
+      seed: () => DraftState(
+        cards: [commonCard, rareCard],
+        selectedCards: const [],
+        status: DraftStateStatus.deckLoaded,
+        firstCardOpacity: 1,
+      ),
+      act: (bloc) {
+        bloc.add(CardSwiped());
+      },
+      verify: (_) {
+        verify(() => audioController.playSfx(Assets.sfx.holoReveal)).called(1);
+      },
+    );
 
     blocTest<DraftBloc, DraftState>(
       'changes opacity order on CardSwipeStarted',
-      build: () => DraftBloc(gameResource: gameResource),
+      build: () => DraftBloc(
+        gameResource: gameResource,
+        audioController: audioController,
+      ),
       seed: () => DraftState(
         cards: cards,
         selectedCards: const [],
@@ -157,8 +303,11 @@ void main() {
     );
 
     blocTest<DraftBloc, DraftState>(
-      'selects a card and changes deck on SelectCard',
-      build: () => DraftBloc(gameResource: gameResource),
+      'selects a card, changes deck on SelectCard, and plays sfx',
+      build: () => DraftBloc(
+        gameResource: gameResource,
+        audioController: audioController,
+      ),
       seed: () => DraftState(
         cards: cards,
         selectedCards: const [],
@@ -179,11 +328,21 @@ void main() {
           firstCardOpacity: 1,
         ),
       ],
+      verify: (_) {
+        final captured =
+            verify(() => audioController.playSfx(captureAny())).captured;
+        final sfx = captured.first;
+        expect(sfx, isA<String>());
+        expect(sfx as String, Assets.sfx.addToHand);
+      },
     );
 
     blocTest<DraftBloc, DraftState>(
       'status is complete when the deck is full',
-      build: () => DraftBloc(gameResource: gameResource),
+      build: () => DraftBloc(
+        gameResource: gameResource,
+        audioController: audioController,
+      ),
       seed: () => DraftState(
         cards: cards,
         selectedCards: [cards[0], cards[1]],
@@ -209,7 +368,10 @@ void main() {
         when(() => gameResource.generateCards(Prompt()))
             .thenThrow(Exception('Error'));
       },
-      build: () => DraftBloc(gameResource: gameResource),
+      build: () => DraftBloc(
+        gameResource: gameResource,
+        audioController: audioController,
+      ),
       act: (bloc) => bloc.add(DeckRequested(Prompt())),
       expect: () => [
         DraftState(
